📈 Implement impossible moves (NEEDS TESTS)
Depth 1 took 0.000166847 seconds to explore 39 positions.
Depth 2 took 0.004045678 seconds to explore 181 positions.
Depth 3 took 0.022155488 seconds to explore 1633 positions.
Depth 4 took 0.12625742 seconds to explore 3927 positions.
Depth 5 took 0.53364503 seconds to explore 46618 positions.
Depth 6 took 4.29317 seconds to explore 193834 positions.
Depth 7 took 17.960434 seconds to explore 1315339 positions.

This slows down our IA for the time being, but we have to take those
impossible moves into account.

+++ src/heuristic.rs
@@ -38,7 +38,7 @@ pub fn sorted_moves(moves: impl Iterator<Item=Move>, favorite_color: Color) -> i
 	})
 }
 
-const fn square_score(x: u64, y: u64) -> u8 {
+const fn square_score(x: i8, y: i8) -> u8 {
 	// Corners, only part of one square.
 	if x == 0 && y == 0 { return 0 }
 	if x == 4 && y == 4 { return 0 }
@@ -51,4 +51,6 @@ const fn square_score(x: u64, y: u64) -> u8 {
 	// Otherwise, the position may have 4 squares, hence
 	// more likely to change the score.
 	return 2
+	// Absolute center does not matter that much, at least
+	// at depth 8.
 }
